feat(harper-lib): move some things into harper_lib and adjust everything accordingly
