+++ harper-core/Cargo.toml
@@ -12,6 +12,7 @@ name = "force-build"
 path = "build.rs"
 
 [dependencies]
+harper-lib = { path = "../harper-lib", version = "0.12.0" }
 harper-dictionary-parsing = { path = "../harper-dictionary-parsing", version = "0.12.0" }
 blanket = "0.4.0"
 fst = "0.4.7"
@@ -35,6 +36,7 @@ rand = "0.8.5"
 
 [build-dependencies]
 fst = "0.4.7"
+harper-lib = { path = "../harper-lib", version = "0.12.0" }
 harper-dictionary-parsing = { path = "../harper-dictionary-parsing", version = "0.12.0" }
 hashbrown = { version = "0.15.0" }
 itertools = "0.13.0"
