+++ harper-core/src/spell/merged_dictionary.rs
@@ -1,5 +1,6 @@
-use itertools::Itertools;
 use std::sync::Arc;
+
+use itertools::Itertools;
 
 use super::dictionary::Dictionary;
 use crate::{CharString, WordMetadata};
