+++ harper-core/src/token.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 use crate::punctuation::Punctuation;
 use crate::Span;
 use crate::{Quote, WordMetadata};
-use harper_dictionary_parsing::word_metadata::{ConjunctionData, NounData};
+use harper_lib::word_metadata::{ConjunctionData, NounData};
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Default)]
 pub struct Token {
