+++ harper-dictionary-parsing/Cargo.toml
@@ -7,8 +7,8 @@ license = "Apache-2.0"
 repository = "https://github.com/elijah-potter/harper"
 
 [dependencies]
+harper-lib = { path = "../harper-lib", version = "0.12.0" }
 hashbrown = { version = "0.15.0", features = ["serde"] }
-is-macro = "0.3.6"
 serde = { version = "1.0.209", features = ["derive"] }
 serde_json = "1.0.132"
 smallvec = { version = "1.13.2", features = ["serde"] }
