+++ harper-dictionary-parsing/src/expansion.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 use super::affix_replacement::{AffixReplacement, HumanReadableAffixReplacement};
 use super::Error;
-use crate::WordMetadata;
+use harper_lib::WordMetadata;
 
 #[derive(Debug, Clone)]
 pub struct Expansion {
