+++ harper-dictionary-parsing/src/lib.rs
@@ -1,22 +1,14 @@
 mod affix_replacement;
 mod attribute_list;
-pub mod char_string;
 mod error;
 mod expansion;
 mod matcher;
-pub mod span;
-pub mod word_list;
-pub mod word_metadata;
+mod word_list;
 
 pub use attribute_list::AttributeList;
 use attribute_list::HumanReadableAttributeList;
-pub use char_string::{CharString, CharStringExt};
 pub use error::Error;
-pub use span::Span;
-pub use word_metadata::WordMetadata;
-
-use self::word_list::parse_word_list;
-pub use self::word_list::MarkedWord;
+pub use word_list::{parse_word_list, MarkedWord};
 
 pub fn parse_default_word_list() -> Result<Vec<MarkedWord>, Error> {
     parse_word_list(include_str!("../dictionary.dict"))
@@ -40,7 +32,7 @@ mod tests {
     use super::word_list::parse_word_list;
     use super::{parse_default_attribute_list, parse_default_word_list};
     use crate::attribute_list::HumanReadableAttributeList;
-    use crate::{CharString, WordMetadata};
+    use harper_lib::{CharString, WordMetadata};
 
     pub const TEST_WORD_LIST: &str = "3\nhello\ntry/B\nwork/AB";
