+++ harper-lib/src/char_string.rs
