+++ harper-lib/src/lib.rs
@@ -1,14 +1,7 @@
-pub fn add(left: u64, right: u64) -> u64 {
-    left + right
-}
+mod char_string;
+mod span;
+pub mod word_metadata;
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn it_works() {
-        let result = add(2, 2);
-        assert_eq!(result, 4);
-    }
-}
+pub use char_string::{CharString, CharStringExt};
+pub use span::Span;
+pub use word_metadata::WordMetadata;
