+++ harper-lib/src/span.rs
