+++ harper-lib/src/word_metadata.rs
