+++ harper-ls/Cargo.toml
@@ -8,6 +8,7 @@ readme = "README.md"
 repository = "https://github.com/elijah-potter/harper"
 
 [dependencies]
+harper-lib = { path = "../harper-lib", version = "0.12.0" }
 harper-core = { path = "../harper-core", version = "0.12.0", features = ["concurrent"] }
 harper-comments = { path = "../harper-comments", version = "0.12.0" }
 harper-html = { path = "../harper-html", version = "0.12.0" }
