+++ harper-ls/src/dictionary_io.rs
@@ -1,6 +1,7 @@
 use std::path::Path;
 
-use harper_core::{Dictionary, FullDictionary, WordMetadata};
+use harper_core::{Dictionary, FullDictionary};
+use harper_lib::WordMetadata;
 use tokio::fs::{self, File};
 use tokio::io::{self, AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt, BufReader, BufWriter};
