+++ harper-tree-sitter/Cargo.toml
@@ -7,5 +7,6 @@ license = "Apache-2.0"
 repository = "https://github.com/elijah-potter/harper"
 
 [dependencies]
+harper-lib = { path = "../harper-lib", version = "0.12.0" }
 harper-core = { path = "../harper-core", version = "0.12.0" }
 tree-sitter = "0.20.10"
