+++ harper-wasm/Cargo.toml
@@ -14,6 +14,7 @@ console_error_panic_hook = "0.1.7"
 tracing = "0.1.40"
 tracing-wasm = "0.2.1"
 wasm-bindgen = "0.2.95"
+harper-lib = { path = "../harper-lib", version = "0.12.0" }
 harper-core = { path = "../harper-core", version = "0.12.0", features = ["concurrent"] }
 once_cell = "1.20.2"
 serde-wasm-bindgen = "0.6.5"
